backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
optuna==3.5.0
xgboost==2.0.2
openai==1.3.7
python-multipart==0.0.6
supabase==2.3.0
psycopg2-binary==2.9.9
pillow==10.1.0